hwc_vsync: Fix Realtime priority usage Commit AOSPA/android_hardware_qcom_display@8490450 was relying on a mix of PRIORITY_URGENT_DISPLAY and PRIORITY_REALTIME. That results in a huge nice value, which makes the task seem important.
However that was missing the real purpose of
the PRIORITY_REALTIME definition.
By summing the values of the two definitions
the Threads controller wasn't able to recognize
ANDROID_PRIORITY_REALTIME which is hardcoded at -21
and apply the important SP_REALTIME scheduling policy
on it.

This commit fixes the flaw by relying on the
PRIORITY_REALTIME definition only.